More unit tests and refactoring Weather Service

+++ warehouse/Warehouse.Models/Interfaces/ICurrentWeather.cs
@@ -2,6 +2,6 @@ namespace Warehouse.Models.Interfaces
 {
     public interface ICurrentWeather
     {
-        double GetTemperatureInCelcius();
+        decimal GetTemperatureInCelcius();
     }
 }
